+++ src/main/java/com/sns255/sns255/domain/email/service/EmailVerificationService.java
@@ -1,0 +1,4 @@
+package com.sns255.sns255.domain.email.service;
+
+public interface EmailVerificationService {
+}
