--- src/main/java/com/sns255/sns255/domain/user/entity/EmailVerificationToken.java
@@ -1,43 +1,0 @@
-package com.sns255.sns255.domain.user.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class EmailVerificationToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String token;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "userId", nullable = false)
-    private User user;
-
-    @Column(nullable = false)
-    private LocalDateTime expiryDate;
-
-    private static final long EXPIRATION_TIME_IN_MINUTES = 15;
-
-    public EmailVerificationToken(String token, User user) {
-        this.token = token;
-        this.user = user;
-        this.expiryDate = calculateExpiryDate();
-    }
-
-    private LocalDateTime calculateExpiryDate() {
-        return LocalDateTime.now().plusMinutes(EXPIRATION_TIME_IN_MINUTES);
-    }
-
-    public boolean isExpired() {
-        return LocalDateTime.now().isAfter(this.expiryDate);
-    }
-}
