+++ src/main/java/com/sns255/sns255/domain/user/repository/EmailVerificationTokenRepository.java
@@ -1,6 +1,5 @@
 package com.sns255.sns255.domain.user.repository;
 
-import com.sns255.sns255.domain.user.entity.EmailVerificationToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
