--- src/main/java/com/sns255/sns255/domain/user/service/EmailVerificationService.java
@@ -1,8 +1,0 @@
-package com.sns255.sns255.domain.user.service;
-
-import com.sns255.sns255.domain.user.entity.User;
-
-public interface EmailVerificationService {
-    String createVerificationToken(User user);
-    void verifyToken(String token);
-}
